refactor: consolidate build scripts and use Bun.watch
